# swagger.yaml
openapi: 3.0.0
paths:
...
  /contracts:
    get:
      tags:
        - contracts
      description: Returns all contracts.
      security:
        - bearerAuth: []
      summary: Returns all contracts.
      parameters:
        - in: query
          name: pageIndex
          schema:
            type: integer
          description: >-
            The number of pages to skip before starting to collect the result
            set.
          examples:
            pageIndex:
              summary: Example of a pageIndex.
              value: 0
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return.
          examples:
            pageSize:
              summary: Example of a pageSize.
              value: 100
      responses:
        '200':
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contract'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - contracts
      description: >-
        Creates new contract. Only users with the Standard Registry role are
        allowed to make the request.
      security:
        - bearerAuth: []
      summary: Create new contract.
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
            example:
              description: Smart Contract Description
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '422':
          description: Unprocessable Entity.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/import:
    post:
      tags:
        - contracts
      description: >-
        Creates new contract. Only users with the Standard Registry role are
        allowed to make the request.
      security:
        - bearerAuth: []
      summary: Import new contract.
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                contractId:
                  type: string
                description:
                  type: string
            example:
              contractId: 0.0.49176956
              description: Smart Contract Description
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/{contractId}/user:
    post:
      tags:
        - contracts
      description: >-
        Add new contract user. Only users with the Standard Registry role are
        allowed to make the request.
      security:
        - bearerAuth: []
      summary: Add new contract user.
      parameters:
        - in: path
          name: contractId
          schema:
            type: string
          description: Contract identifier
          required: true
          example: 0.0.49176956
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
            example:
              userId: 0.0.49176949
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/{contractId}/status:
    post:
      tags:
        - contracts
      description: >-
        Update contract status. Only users with the Standard Registry role are
        allowed to make the request.
      security:
        - bearerAuth: []
      summary: Update contract status.
      parameters:
        - in: path
          name: contractId
          schema:
            type: string
          description: Contract identifier
          required: true
          example: 0.0.49176956
      responses:
        '200':
          description: >-
            Successful operation. Returns true if you are added to contract else
            false.
          content:
            application/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/pair:
    get:
      tags:
        - contracts
      description: Returns all contracts pairs.
      security:
        - bearerAuth: []
      summary: Returns all contracts pairs.
      parameters:
        - in: query
          name: baseTokenId
          schema:
            type: string
          description: Base token identifier
          example: 0.0.49179655
        - in: query
          name: oppositeTokenId
          schema:
            type: string
          description: Opposite token identifier
          example: 0.0.49179656
      responses:
        '200':
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    baseTokenRate:
                      type: integer
                    oppositeTokenRate:
                      type: integer
                    contractId:
                      type: string
                    description:
                      type: string
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/{contractId}/pair:
    post:
      tags:
        - contracts
      description: >-
        Creates new contract pair. Only users with the Standard Registry role
        are allowed to make the request.
      security:
        - bearerAuth: []
      summary: Creates new contract pair.
      parameters:
        - in: path
          name: contractId
          schema:
            type: string
          description: Contract identifier
          required: true
          example: 0.0.49176956
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                baseTokenId:
                  type: string
                oppositeTokenId:
                  type: string
                baseTokenCount:
                  type: integer
                oppositeTokenCount:
                  type: integer
            example:
              baseTokenId: 0.0.49179655
              oppositeTokenId: 0.0.49179656
              baseTokenCount: 1
              oppositeTokenCount: 3
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/retire/request:
    get:
      tags:
        - contracts
      description: Returns all contract requests.
      security:
        - bearerAuth: []
      summary: Returns all contract requests.
      parameters:
        - in: query
          name: contractId
          schema:
            type: string
          description: Contract Identifier.
          example: 0.0.49176956
        - in: query
          name: pageIndex
          schema:
            type: integer
          description: >-
            The number of pages to skip before starting to collect the result
            set.
          examples:
            pageIndex:
              summary: Example of a pageIndex.
              value: 0
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return.
          examples:
            pageSize:
              summary: Example of a pageSize.
              value: 100
      responses:
        '200':
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetireRequest'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - contracts
      description: Cancel contract requests.
      security:
        - bearerAuth: []
      summary: Cancel contract requests.
      parameters:
        - in: query
          name: requestId
          schema:
            type: string
          description: Request Identifier.
          required: true
          example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/{contractId}/retire/request:
    post:
      tags:
        - contracts
      description: Creates new contract retire request.
      security:
        - bearerAuth: []
      summary: Creates new contract retire request.
      parameters:
        - in: path
          name: contractId
          schema:
            type: string
          description: Contract identifier
          required: true
          example: 0.0.49176956
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                baseTokenId:
                  type: string
                oppositeTokenId:
                  type: string
                baseTokenCount:
                  type: integer
                oppositeTokenCount:
                  type: integer
                baseTokenSerials:
                  type: array
                  items:
                    type: integer
                oppositeTokenSerials:
                  type: array
                  items:
                    type: integer
            example:
              baseTokenId: 0.0.49179655
              oppositeTokenId: 0.0.49179656
              baseTokenCount: 2
              oppositeTokenCount: 6
              baseTokenSerials: []
              oppositeTokenSerials: []
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetireRequest'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/retire:
    post:
      tags:
        - contracts
      description: >-
        Retire tokens. Only users with the Standard Registry role are allowed to
        make the request.
      security:
        - bearerAuth: []
      summary: Retire tokens.
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                requestId:
                  type: string
            example:
              requestId: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
info:
  title: Guardian
  description: >-
    The Guardian is a modular open-source solution that includes best-in-class
    identity management and decentralized ledger technology (DLT) libraries. At
    the heart of the Guardian solution is a sophisticated Policy Workflow Engine
    (PWE) that enables applications to offer a requirements-based tokenization
    implementation.
  version: 2.16.0
  contact:
    name: API developer
    url: https://envisionblockchain.com
    email: info@envisionblockchain.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: /api/v1
    description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
    Contract:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        description:
          type: string
        owner:
          type: string
        isOwnerCreator:
          type: string
        status:
          type: string
    RetireRequest:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        baseTokenId:
          type: string
        owner:
          type: string
        oppositeTokenId:
          type: string
        baseTokenCount:
          type: number
        oppositeTokenCount:
          type: number
...